+++ HISTORY.rst
@@ -6,6 +6,33 @@ History
 This section summarizes all my steps on the way to a working micc,
 including dead-ends.
 
+v0.5.13 (2019-09-19)
+====================
+improving tests. I am still rather uncomfortable with testing cli's. Most of the code is tested by 
+manual inspection. This unlucky patch number is for improving this situation.
+
+* step 1: study the click documentation on testing.
+  Interesting topic: File system isolation, runs a scenario in an empty directory,
+  e.g. /private/var/folders/rt/7h5lk6c955db20y1rzf1rjz00000gn/T/tmpa12gc_p9
+  Good idea but the location of that directory is a bit hard to trace. (Note, that
+  a temporary directory inside micc doesn't work, because micc refuses to create a 
+  project inside another project. 
+  
+refactoring: still not happy with the ``--simple`` argument for micc 
+
+* how does python refer to its project structure? ``module.py`` vs ``module/__init__.py``
+
+  * ``module.py`` is called a *module*
+  * ``module/__init__.py`` is called a *package*
+  
+  Thus, our *modules* may be *modules* or sub-level *packages*, and
+  our *package* is actually the top-level package, which may contain
+  other lower-level packages, as well as modules.
+ 
+v0.5.12 (2019-09-19)
+====================
+Add logging to micc commands as we tried out in micc build
+
 v0.5.11 (2019-09-18)
 ====================
 #2 [feature] simple python project